organize experience file structure
